chore: Improvements for shutdown and add a test
